Data.Conduit.Chan.Unagi

Description

This module dumbly wraps up blocking Unagi channels into conduit sources and sinks.

Documentation

sourceChan :: (MonadIO m, Readable chan, ReadableConstraint chan a) => chan a -> Source m a

Data arriving on the out-channel is passed to the source

sinkChan :: (MonadIO m, Writeable chan, WriteableConstraint chan a) => chan a -> Sink a m ()

Data arriving on the sink is passed to the in-channel