Data.Conduit.Chan.Unagi.Types

Description

This module dumbly rounds up Unagi channels into a Readable and a Writeable typeclass, to be used by Data.Conduit.Chan.Unagi.

Documentation

class Readable chan where

A typeclass for readable, blocking Unagi channels

Associated Types

type ReadableConstraint chan a :: Constraint

Methods

readChan :: ReadableConstraint chan a => chan a -> IO a

class Writeable chan where

A typeclass for writeable, blocking Unagi channels

Associated Types

type WriteableConstraint chan a :: Constraint

Methods

writeChan :: WriteableConstraint chan a => chan a -> a -> IO ()